split_manufacturing,xg_sequential_planning

+++ split_manufacturing/models/manufacturing_split.py
@@ -5,6 +5,7 @@ from odoo import api, fields, models, Command
 from odoo.tools import float_round, float_compare
 
 
+
 class MrpProductionSplitMulti(models.TransientModel):
     _name = 'mrp.production.split.multi'
     _description = "Wizard to Split Multiple Productions"
@@ -28,6 +29,7 @@ class MrpProductionSplit(models.TransientModel):
         'mrp.production.split.line', 'mrp_production_split_id',
         'Split Details', compute="_compute_details", store=True, readonly=False)
     valid_details = fields.Boolean("Valid", compute="_compute_valid_details")
+    split_sequence = fields.Float(related='production_id.split_sequence')
 
 
     @api.depends('production_detailed_vals_ids')
@@ -39,10 +41,6 @@ class MrpProductionSplit(models.TransientModel):
     def _compute_details(self):
         for wizard in self:
             commands = [Command.clear()]
-            # print(commands),
-            # print(Command,2222222222222)
-            # import pdb;
-            # pdb.set_trace()
             if wizard.counter < 1 or not wizard.production_id:
                 wizard.production_detailed_vals_ids = commands
                 continue
@@ -68,7 +66,7 @@ class MrpProductionSplit(models.TransientModel):
             }))
             wizard.production_detailed_vals_ids = commands
 
-       
+        
 
     @api.depends('production_detailed_vals_ids')
     def _compute_valid_details(self):
@@ -78,23 +76,82 @@ class MrpProductionSplit(models.TransientModel):
                 wizard.valid_details = float_compare(wizard.product_qty, sum(wizard.production_detailed_vals_ids.mapped('quantity')), precision_rounding=wizard.product_uom_id.rounding) == 0
 
     def action_split(self):
-        # import pdb;
-        # pdb.set_trace()
-        productions = self.production_id._split_productions({self.production_id: [detail.quantity for detail in self.production_detailed_vals_ids]})
-        # print(productions,"PRODUCTIONS")
-        # productions[0].copy()
-        # import pdb;
-        # pdb.set_treace()
-        # print(productions,"PRODUCTIONS")
-        for production, detail in zip(productions, self.production_detailed_vals_ids):
-            production.user_id = detail.user_id
-            production.date_planned_start = detail.date
-        if self.production_split_multi_id:
-            saved_production_split_multi_id = self.production_split_multi_id.id
-            self.production_split_multi_id.production_ids = [Command.unlink(self.id)]
-            action = self.env['ir.actions.actions']._for_xml_id('split_manufacturing.action_mrp_production_split_multi')
-            action['res_id'] = saved_production_split_multi_id
-            return action
+        productions = self.production_id._split_productions({self.production_id : [detail.quantity for detail in self.production_detailed_vals_ids]})
+        if productions:
+            # for origin in productions:
+            #     # origin.origin = productions[0].origin + ',' + ",".join(sorted([production.name for production in productions[1:]]))
+            #     origin.origin = ",".join(sorted([production.name for production in productions]))
+
+            # saleorder = set([production.so_origin for production in productions])
+            # if len(saleorder) == 2:
+            #     so_origin = list(saleorder)[1]
+            # else:
+            #     so_origin = saleorder.pop()       
+            
+            # # import pdb;
+            # # pdb.set_trace()
+
+
+            # productcolour = set([production.product_colour for production in productions])
+            # if len(productcolour) == 2:
+            #     product_colour = list(productcolour)[1]
+            # else:
+            #     product_colour = productcolour.pop()
+
+            
+            # morign = productions[0].origin.split(',')
+            # if morign:
+            #     mo_orign = morign[0]
+            # else:
+            #     mo_orign = False
+
+            # if so_origin == False:
+            #     search_mo_id = self.env['mrp.production'].search([('origin','=',mo_orign)]).filtered(lambda x:x.so_origin != False and x.product_colour != False)
+            #     if len(search_mo_id)>1:
+            #         so_origin = search_mo_id.so_origin
+            #     else:
+            #         for search_mo in search_mo_id:
+            #             if search_mo.so_origin:
+            #                 so_origin = search_mo_id.so_origin
+
+            
+            # if product_colour == False:
+            #     search_mo_id = self.env['mrp.production'].search([('origin','=',mo_orign)]).filtered(lambda x:x.so_origin != False and x.product_colour != False)
+            #     if len(search_mo_id)>1:
+            #         product_colour = search_mo_id.product_colour
+            #     else:
+            #         for search_mo in search_mo_id:
+            #             if search_mo.product_colour:
+            #                 product_colour = search_mo_id.product_colour
+
+
+            # if product_colour == False:
+            #     product_colour = self.production_detailed_vals_ids[0].product_colour
+
+            # if so_origin == False:
+            #     so_origin = self.production_detailed_vals_ids[0].so_origin
+            product_colour = self.production_detailed_vals_ids[0].product_colour
+            so_origin = self.production_detailed_vals_ids[0].so_origin
+            
+            value = 0
+            for mos in productions:
+                mos.split_sequence = round(self.production_detailed_vals_ids[value].serial_no,2)
+                value += 1
+                if so_origin:
+                    mos.so_origin = so_origin
+                if product_colour:
+                    mos.product_colour = product_colour
+            # productions[0].write({'is_splited' : True})
+            for production, detail in zip(productions, self.production_detailed_vals_ids):
+                production.user_id = detail.user_id
+                production.date_planned_start = detail.date
+            if self.production_split_multi_id:
+                saved_production_split_multi_id = self.production_split_multi_id.id
+                self.production_split_multi_id.production_ids = [Command.unlink(self.id)]
+                action = self.env['ir.actions.actions']._for_xml_id('split_manufacturing.action_mrp_production_split_multi')
+                action['res_id'] = saved_production_split_multi_id
+                return action
+
 
     def action_prepare_split(self):
         action = self.env['ir.actions.actions']._for_xml_id('split_manufacturing.action_mrp_production_split')
@@ -122,3 +179,18 @@ class MrpProductionSplitLine(models.TransientModel):
     date = fields.Datetime('Schedule Date')
     product_colour = fields.Char()
     so_origin = fields.Char()
+    
+    serial_no = fields.Float(string='#',compute='_compute_sl')
+    
+    @api.depends('quantity','user_id')
+    def _compute_sl(self):
+        for order in self.mapped('mrp_production_split_id'):
+            number = order.split_sequence
+            for line in order.production_detailed_vals_ids:
+                if line == order.production_detailed_vals_ids[0]:
+                    line.serial_no = number
+                else:
+                    number += 0.01
+                    line.serial_no = number
+                    
+
